+++ PZPP_Biblioteka/GatunekViewModel.cs
@@ -1,14 +1,15 @@
-﻿using System;
+﻿using Microsoft.EntityFrameworkCore;
+using PZPP_Biblioteka;
+using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Input;
-using Microsoft.EntityFrameworkCore;
-using System.Runtime.CompilerServices;
-using PZPP_Biblioteka;
 
 namespace PZPP_Biblioteka
 {
@@ -96,9 +97,18 @@ namespace PZPP_Biblioteka
         {
             if (SelectedGatunek == null) return;
 
-            _context.GatunkiKsiążek.Remove(SelectedGatunek);
-            _context.SaveChanges();
-            OdswiezGatunek();
+            var wynik = MessageBox.Show(
+                        $"Czy na pewno chcesz usunąć gatunek:\n{SelectedGatunek.Nazwa}?",
+                        "Potwierdzenie usunięcia",
+                        MessageBoxButton.YesNo,
+                        MessageBoxImage.Warning);
+
+            if (wynik == MessageBoxResult.Yes)
+            {
+                _context.GatunkiKsiążek.Remove(SelectedGatunek);
+                _context.SaveChanges();
+                OdswiezGatunek();
+            }
         }
 
         public void OdswiezGatunek()
