+++ PZPP_Biblioteka/KsiążkaViewModel.cs
@@ -16,12 +16,15 @@ namespace PZPP_Biblioteka
     public class KsiążkaViewModel : INotifyPropertyChanged
     {
         private readonly Biblioteka _context;
+        private readonly ApiService _api;
         public ObservableCollection<Książka> Książki { get; set; }
+    
 
         public ICommand PokazDodajKsiążkaCommand { get; }
         public ICommand EdytujKsiążkaCommand { get; }
         public ICommand UsunKsiążkaCommand { get; }
         public ICommand ZapiszKsiążkaCommand { get; }
+        public ICommand SortujCommand { get; }
         public event Action ZamknijOkno;
 
         private Książka _selectedKsiążka;
@@ -47,15 +50,62 @@ namespace PZPP_Biblioteka
                 OnPropertyChanged();
             }
         }
+        private async void ZaladujDostepnosc()
+        {
+            var lista = Książki.ToList();
+
+            foreach (var k in lista)
+            {
+                var wynik = await _api.PobierzDostepnosc(k.Tytuł);
+
+                Application.Current.Dispatcher.Invoke(() =>
+                {
+                    k.DostepnoscWBibliotece2 = wynik;
+                });
+            }
+
+            OnPropertyChanged(nameof(Książki));
+        }
+
+        private int _dostepnosc;
+        public int DostepnoscWBibliotece2
+        {
+            get => _dostepnosc;
+            set
+            {
+                _dostepnosc = value;
+                OnPropertyChanged();
+            }
+        }
+        private string _wyszukaj;
+
+        public string Wyszukaj
+        {
+            get => _wyszukaj;
+            set
+            {
+                _wyszukaj = value;
+                OnPropertyChanged();
+                FiltrujKsiążki();
+                
+            }
+        }
 
         public KsiążkaViewModel(Biblioteka context)
         {
             _context = context;
-            Książki = new ObservableCollection<Książka>(_context.Książki.ToList());
+            _api = new ApiService();
+            Książki = new ObservableCollection<Książka>(_context.Książki
+                            .Include(k => k.Autor)
+                            .Include(k => k.GatunekKsiążki)
+                            .ToList());
             PokazDodajKsiążkaCommand = new RelayCommand(PokazDodajKsiążka);
             EdytujKsiążkaCommand = new RelayCommand(EdytujKsiążka, _ => SelectedKsiążka != null);
             UsunKsiążkaCommand = new RelayCommand(UsunKsiążka, _ => SelectedKsiążka != null);
             ZapiszKsiążkaCommand = new RelayCommand(ZapiszKsiążka, _ => SelectedKsiążka != null);
+            SortujCommand = new RelayCommand(_ => SortujAlfabetycznie());
+
+            ZaladujDostepnosc();
         }
 
         private void PokazDodajKsiążka(object obj)
@@ -91,6 +141,12 @@ namespace PZPP_Biblioteka
                 _context.SaveChanges();
             }
 
+            if (string.IsNullOrWhiteSpace(SelectedKsiążka.Tytuł))
+            {
+                MessageBox.Show("Tytuł książki jest wymagany!");
+                return;
+            }
+
             ZamknijOkno?.Invoke();
         }
 
@@ -98,17 +154,69 @@ namespace PZPP_Biblioteka
         {
             if (SelectedKsiążka == null) return;
 
-            _context.Książki.Remove(SelectedKsiążka);
-            _context.SaveChanges();
-            OdswiezKsiążki();
+            var wynik = MessageBox.Show($"Czy na pewno chcesz usunąć książkę:\n{SelectedKsiążka.Tytuł}?","Potwierdzenie usunięcia",MessageBoxButton.YesNo,MessageBoxImage.Warning);
+
+            if (wynik == MessageBoxResult.Yes)
+            {
+                _context.Książki.Remove(SelectedKsiążka);
+                _context.SaveChanges();
+                OdswiezKsiążki();
+            }
         }
 
         public void OdswiezKsiążki()
         {
             Książki.Clear();
-            foreach (var ksiazka in _context.Książki.Include(k => k.GatunekKsiążki).ToList())
+            foreach (var ksiazka in _context.Książki.Include(k => k.GatunekKsiążki).Include(k => k.Autor).ToList())
             {
                 Książki.Add(ksiazka);
+            }
+        }
+
+        private void FiltrujKsiążki()
+        {
+            if (string.IsNullOrWhiteSpace(Wyszukaj))
+            {
+                OdswiezKsiążki();
+                return;
+            }
+
+            Książki.Clear();
+
+            var tekst = Wyszukaj?.ToLower() ?? "";
+
+            var wynik = _context.Książki
+                .Include(k => k.Autor)
+                .Include(k => k.GatunekKsiążki)
+                .Where(k =>
+
+                    k.Tytuł.ToLower().Contains(tekst)
+
+                    || k.Autor.Imię.ToLower().Contains(tekst)
+
+                    || k.Autor.Nazwisko.ToLower().Contains(tekst)
+
+                    || k.GatunekKsiążki.Nazwa.ToLower().Contains(tekst)
+                )
+                .ToList();
+
+            foreach (var k in wynik)
+            {
+                Książki.Add(k);
+            }
+        }
+
+        private void SortujAlfabetycznie()
+        {
+            var posortowane = Książki
+                .OrderBy(k => k.Tytuł)
+                .ToList();
+
+            Książki.Clear();
+
+            foreach (var k in posortowane)
+            {
+                Książki.Add(k);
             }
         }
